requirements.txt
openai>=1.99.0
python-dotenv>=1.0.1
python-telegram-bot>=21.0
slack-bolt>=1.22.0
